move StartPage to folder and add CSS Module
