+++ src/pages/Start/StartPage.tsx
@@ -1,13 +1,14 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import Header from '../components/Header/Header';
-import Button from '../components/Button/Button';
-import Card from '../components/Card/Card';
-import SettingsForm from '../components/SettingsForm/SettingsForm';
-import Modal from '../components/Modal/Modal';
-import { useQuizData } from '../hooks/useQuizData';
-import { useGameSettings } from '../context/GameSettingsContext';
-import type { GameSettings } from '../types';
+import Header from '../../components/Header/Header';
+import Button from '../../components/Button/Button';
+import Card from '../../components/Card/Card';
+import SettingsForm from '../../components/SettingsForm/SettingsForm';
+import Modal from '../../components/Modal/Modal';
+import { useQuizData } from '../../hooks/useQuizData';
+import { useGameSettings } from '../../context/GameSettingsContext';
+import type { GameSettings } from '../../types';
+import styles from './StartPage.module.css';
 
 const StartPage: React.FC = () => {
   const navigate = useNavigate();
@@ -17,9 +18,8 @@ const StartPage: React.FC = () => {
 
   const handleStart = () => {
     if (questions.length > 0) {
-      // Зберігаємо питання та ID користувача в sessionStorage
       sessionStorage.setItem('quizQuestions', JSON.stringify(questions));
-      sessionStorage.setItem('currentUserId', '1'); // Простий ID для прикладу
+      sessionStorage.setItem('currentUserId', '1');
       navigate('/game');
     }
   };
@@ -34,29 +34,28 @@ const StartPage: React.FC = () => {
   };
 
   const handleUserProfile = () => {
-    // Отримуємо ID користувача з sessionStorage або використовуємо дефолтний
     const userId = sessionStorage.getItem('currentUserId') || '1';
     navigate(`/user/${userId}`);
   };
 
   return (
-    <div className="page start-page">
+    <div className="page">
       <Header />
       
-      <Card className="start-page__card">
-        <div className="start-page__content">
-          <h2>Ласкаво просимо до Кіно-Вікторини!</h2>
+      <Card size="large" className={styles.card}>
+        <div className={styles.content}>
+          <h2>Ласкаво просимо до Кіно-Вікторини! 🎬</h2>
 
-          <div className="current-settings">
+          <div className={styles.currentSettings}>
             <h4>Поточні налаштування:</h4>
-            <div className="settings-preview">
+            <div className={styles.settingsPreview}>
               <span>Питань: <strong>{settings.count}</strong></span>
               <span>Складність: <strong>{getDifficultyLabel(settings.difficulty)}</strong></span>
               <span>Час: <strong>{settings.timerDuration}с</strong></span>
             </div>
           </div>
 
-          <div className="rules">
+          <div className={styles.rules}>
             <h3>Правила гри:</h3>
             <ul>
               <li>Відповідайте на запитання про фільми</li>
@@ -67,13 +66,13 @@ const StartPage: React.FC = () => {
           </div>
 
           {loading && (
-            <div className="loading-state">
+            <div className={styles.loadingState}>
               <p>Завантаження питань...</p>
             </div>
           )}
 
           {error && (
-            <div className="error-state">
+            <div className={styles.errorState}>
               <p>Помилка: {error.message}</p>
               <Button onClick={() => reload()}>
                 Спробувати знову
@@ -82,7 +81,7 @@ const StartPage: React.FC = () => {
           )}
 
           {!loading && !error && questions.length === 0 && (
-            <div className="empty-state">
+            <div className={styles.emptyState}>
               <p>Не знайдено питань за обраними критеріями</p>
               <Button onClick={() => reload()}>
                 Оновити
@@ -90,7 +89,7 @@ const StartPage: React.FC = () => {
             </div>
           )}
 
-          <div className="start-actions">
+          <div className={styles.actions}>
             <Button onClick={handleShowSettings} variant="secondary">
               Налаштування
             </Button>
